Extracted the common features both login and signup forms to a render-Prop to further reduce code repetition
