+++ client/src/pages/Accounts/Signup/SignUp.tsx
@@ -1,38 +1,10 @@
-import { FormikHelpers } from 'formik';
-import AccountWrapper from '../../../components/AccountWrapper/AccountWrapper';
-import register from '../../../helpers/APICalls/register';
+import UserAccountWrapper from '../../../components/UserAccountWrapper/UserAccountWrapper';
 import SignUpForm from '../AuthForm/SignupForm/SignUpForm';
-import { useAuth } from '../../../context/useAuthContext';
-import { useSnackBar } from '../../../context/useSnackbarContext';
 
 export default function Register(): JSX.Element {
-  const { updateLoginContext } = useAuth();
-  const { updateSnackBarMessage } = useSnackBar();
-
-  const handleSubmit = (
-    { username, email, password }: { email: string; password: string; username: string },
-    { setSubmitting }: FormikHelpers<{ email: string; password: string; username: string }>,
-  ) => {
-    register(username, email, password).then((data) => {
-      if (data.error) {
-        console.error({ error: data.error.message });
-        setSubmitting(false);
-        updateSnackBarMessage(data.error.message);
-      } else if (data.success) {
-        updateLoginContext(data.success);
-      } else {
-        // should not get here from backend but this catch is for an unknown issue
-        console.error({ data });
-
-        setSubmitting(false);
-        updateSnackBarMessage('An unexpected error occurred. Please try again');
-      }
-    });
-  };
-
   return (
-    <AccountWrapper>
-      <SignUpForm handleSubmit={handleSubmit} />
-    </AccountWrapper>
+    <UserAccountWrapper>
+      {(handleLogin, handleRegister) => <SignUpForm handleSubmit={handleRegister} />}
+    </UserAccountWrapper>
   );
 }
